refactor(server): identity is ordinary code over the claimed auth, not a registry entry
Registry audit, applying the rule the drain simplification set: the registry
carries resources (live objects that outlive swaps) and capabilities (process
singletons only the runtime can hand over) — anything derivable from an
already-claimed capability is neither, and being registered anyway makes it a
duplicate channel pretending to be a resource.

runtime:identity was exactly that. Its stated reason — "don't make the bundle
re-implement cookie parsing and session lookup" — stopped being true the day
the business surface moved into the bundle: auth/middleware.ts rides in the
bundle and authenticates every business request with the same SESSION_COOKIE
and the same claimed AuthService. The published resolver was a second copy of
"cookie → user" flowing through a registry id, with authenticateWithMeta
already a handshake-verified member of the auth capability.

identityFrom now BUILDS the resolver from caps.authService (readSessionCookie
moves next to it; app.ts loses the registration and the helper). A bare
kernel has no auth to claim and authenticates nobody — terminals stay
fail-closed there, the same behavior the registry fallback had. One
compatibility note: a previously committed bundle restored on this runtime
still claims the dead id and gets nobody — terminals 401 until the next
push. Fail-closed, pre-release, accepted.

The rest of the audit (kept, with reasons): terminal:* are the archetypal
resources; runtime:{config,db,auth,channels,proxy,hmr,desktop} are the
capability handshake; runtime:interfaces / resource-interfaces / bare-kernel
are contract declarations for which the registry is the only cross-unit or
cross-generation channel; runtime:overrides is the test-only injection
channel (live functions cannot ride the parked doc); hmr:assets is the
host→bundle pointer to per-push native modules; platform:current is the one
reverse-direction capability — the App is rebuilt every generation, and its
four runtime readers (channel sweep, auth provisioning, bootAppDeps,
shutdown) need a late-binding entry point, which after the drain
simplification costs exactly one overwritten key and zero releases.

Server suite 813/813, typecheck and Prettier clean.
